+++ apps/server/src/modules/news/controller/news.controller.ts
@@ -1,23 +1,27 @@
-import { Controller, Get, Post, Body, Param, UseInterceptors, ClassSerializerInterceptor, Query } from '@nestjs/common';
-import { NewsUc } from '../uc/news.uc';
+import { Controller, Get, Post, Body, Param, Query, Patch, Delete } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
+import { ParseObjectIdPipe } from '@shared/controller/pipe';
+import { PaginationQuery } from '@shared/controller';
+import { PaginationResponse } from '@shared/controller/dto/pagination.response';
+import { NewsUc } from '../uc/news.uc';
 import { Authenticate, CurrentUser } from '../../authentication/decorator/auth.decorator';
 import { ICurrentUser } from '../../authentication/interface/jwt-payload';
-import { ParseObjectIdPipe } from '../../../shared/core/pipe/parse-object-id.pipe';
-import { CreateNewsRequestDto, NewsResponseDto } from './dto';
-import { PaginationQueryDto } from '../../../shared/core/controller/dto/pagination.query.dto';
+import { CreateNewsParams, NewsFilterQuery, NewsResponse, UpdateNewsParams } from './dto';
 import { NewsMapper } from '../mapper/news.mapper';
 
 @ApiTags('News')
 @Authenticate('jwt')
 @Controller('news')
-@UseInterceptors(ClassSerializerInterceptor) // works only for class instances, default object are not covered!
 export class NewsController {
 	constructor(private readonly newsUc: NewsUc) {}
 
 	@Post()
-	async create(@Body() createNewsDto: CreateNewsRequestDto): Promise<NewsResponseDto> {
-		const news = await this.newsUc.create(NewsMapper.mapCreateNewsToDomain(createNewsDto));
+	async create(@CurrentUser() currentUser: ICurrentUser, @Body() params: CreateNewsParams): Promise<NewsResponse> {
+		const news = await this.newsUc.create(
+			currentUser.userId,
+			currentUser.schoolId,
+			NewsMapper.mapCreateNewsToDomain(params)
+		);
 		const dto = NewsMapper.mapToResponse(news);
 		return dto;
 	}
@@ -25,11 +29,16 @@ export class NewsController {
 	@Get()
 	async findAll(
 		@CurrentUser() currentUser: ICurrentUser,
-		@Query() pagination: PaginationQueryDto
-	): Promise<NewsResponseDto[]> {
-		const newsList = await this.newsUc.findAllForUser(currentUser.userId, pagination);
+		@Query() scope: NewsFilterQuery,
+		@Query() pagination: PaginationQuery
+	): Promise<PaginationResponse<NewsResponse[]>> {
+		const [newsList, count] = await this.newsUc.findAllForUser(
+			currentUser.userId,
+			NewsMapper.mapNewsScopeToDomain(scope),
+			pagination
+		);
 		const dtoList = newsList.map((news) => NewsMapper.mapToResponse(news));
-		return dtoList;
+		return new PaginationResponse(dtoList, count);
 	}
 
 	/** Retrieve a specific news entry by id. A user may only read news of scopes he has the read permission. The news entity has school and user names populated. */
@@ -38,23 +47,29 @@ export class NewsController {
 		// A parameter pipe like ParseObjectIdPipe gives us the guarantee of typesafety for @Param
 		@Param('id', ParseObjectIdPipe) newsId: string,
 		@CurrentUser() currentUser: ICurrentUser
-	): Promise<NewsResponseDto> {
+	): Promise<NewsResponse> {
 		const news = await this.newsUc.findOneByIdForUser(newsId, currentUser.userId);
 		const dto = NewsMapper.mapToResponse(news);
 		return dto;
 	}
 
-	// @Patch(':id')
-	// update(
-	// 	@Param('id', ParseObjectIdPipe)
-	// 	newsId: ObjectId,
-	// 	@Body() updateNewsDto: UpdateNewsDto
-	// ) {
-	// 	return this.newsService.update(newsId, updateNewsDto);
-	// }
+	@Patch(':id')
+	async update(
+		@Param('id', ParseObjectIdPipe) newsId: string,
+		@CurrentUser() currentUser: ICurrentUser,
+		@Body() params: UpdateNewsParams
+	): Promise<NewsResponse> {
+		const news = await this.newsUc.update(newsId, currentUser.userId, NewsMapper.mapUpdateNewsToDomain(params));
+		const dto = NewsMapper.mapToResponse(news);
+		return dto;
+	}
 
-	// @Delete(':id')
-	// remove(@Param('id') id: string): Promise<string> {
-	// 	return this.newsService.remove(id);
-	// }
+	@Delete(':id')
+	async delete(
+		@Param('id', ParseObjectIdPipe) newsId: string,
+		@CurrentUser() currentUser: ICurrentUser
+	): Promise<string> {
+		const deletedId = await this.newsUc.delete(newsId, currentUser.userId);
+		return deletedId;
+	}
 }