+++ apps/server/src/modules/news/mapper/user-info.mapper.ts
@@ -1,15 +1,12 @@
 import { UserInfo } from '../entity';
-import { UserInfoResponseDto } from '../controller/dto';
+import { UserInfoResponse } from '../controller/dto';
 
 export class UserInfoMapper {
-	static mapToResponse(userInfo: UserInfo): UserInfoResponseDto {
-		if (userInfo != null) {
-			const dto = new UserInfoResponseDto();
-			dto.id = userInfo.id;
-			dto.firstName = userInfo.firstName;
-			dto.lastName = userInfo.lastName;
-			return dto;
-		}
-		return null;
+	static mapToResponse(userInfo: UserInfo): UserInfoResponse {
+		const dto = new UserInfoResponse();
+		dto.id = userInfo.id;
+		dto.firstName = userInfo.firstName;
+		dto.lastName = userInfo.lastName;
+		return dto;
 	}
 }