+++ apps/server/src/modules/news/uc/create-news.dto.ts
@@ -1,5 +1,0 @@
-export interface ICreateNewsDto {
-	title: string;
-	content: string;
-	displayAt: Date;
-}