+++ apps/server/src/modules/task/controller/dto/index.ts
@@ -1,1 +1,1 @@
-export * from './task-response.dto';
+export * from './task.response';