+++ apps/server/src/shared/common/interceptor/constants.ts
@@ -1,0 +1,5 @@
+import config = require('@hpi-schul-cloud/commons');
+
+const { Configuration } = config;
+
+export const REQUEST_TIMEOUT = Configuration.get('REQUEST_TIMEOUT') as number;