+++ apps/server/src/shared/repo/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Shared repository models, consists of entities and interfaces.
+ * Repositories must provide their own interfaces.
+ * It is forbidden to reuse existing ones from different layers.
+ */
+
+export * from './query';