+++ apps/server/src/shared/repo/query/empty-result.query.ts
@@ -1,0 +1,5 @@
+/**
+ * When this query is added ($and) to an existing query,
+ * it should ensure an empty result
+ */
+export const EmptyResultQuery = { $and: [{ _id: false }] };