+++ apps/server/test/app.e2e-spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
 import * as request from 'supertest';
-import { ServerModule } from './../src/server.module';
+import { ServerModule } from '@src/server.module';
 
 describe('ServerController (e2e)', () => {
 	let app: INestApplication;
@@ -15,8 +15,11 @@ describe('ServerController (e2e)', () => {
 		await app.init();
 	});
 
-	// todo: fix open handle
+	afterEach(async () => {
+		await app.close();
+	});
+
 	it('/ (GET)', () => {
-		return request(app.getHttpServer()).get('/').expect(200).expect('Hello World!');
+		return request(app.getHttpServer()).get('/').expect(200).expect('Schulcloud Server API');
 	});
 });